bnb-943 | Remove lines of links in title (#28)
* styling: make links without underline to be closer to without

Our AU links have an underline but in WU this is different. Changed the scrollpage titles to not have an underline

Refs: bnb-943

* debug: try out openNewEmbed and back from SDK

Refs: bnb-943

* debug: fix not being send to overview page because of navigate back is triggered

When selecitng a tile vlaanderen or aalter go get navigated back instantly as this is what the code said. Changed this so it needs to come from a detail page.

Refs: bnb-943

* debug: also check the from route when opening a new embed

Refs: bnb-943

* debug: remove all sdk routing

Ref: bnb-943